+++ src/main/java/org/chunsik/pq/login/dto/MeResponseDto.java
@@ -4,11 +4,11 @@ import lombok.Getter;
 
 @Getter
 public class MeResponseDto {
-    private final Integer id;
+    private final Long id;
     private final String email;
     private final String nickname;
 
-    public MeResponseDto(Integer id, String email, String nickname) {
+    public MeResponseDto(Long id, String email, String nickname) {
         this.id = id;
         this.email = email;
         this.nickname = nickname;
